Spin Class Added
Spin obstacle class has been added to the branch Paul ready to be added
to the master branch.

The class consists of three spinning obstacles.
The obstacles can be stopped by tapping the screen.
A lever has also been added which will be used in future to stop the
spinning obstacles.